Handle the most basic child case

+++ hk2-configuration/persistence/hk2-xml/src/main/java/org/glassfish/hk2/xml/jaxb/internal/BaseHK2JAXBBean.java
@@ -125,7 +125,7 @@ public class BaseHK2JAXBBean {
     }
     
     public long _getPropertyJ(String propName) {
-        return (Integer) beanLikeMap.get(propName);
+        return (Long) beanLikeMap.get(propName);
     }
     
     public double _getPropertyD(String propName) {
